Add option to disable universe logging file
Universe logs quite a big amount of logs to /tmp if
the agent runs long time. There is a risk to drain
out the disk space on /tmp, which might result in
no enough space if the /tmp is mounted on disk; or
run out of memory if it's a tempfs.